python-supabase/requirements.txt
flask==2.2.2
flask-cors==3.0.10
python-dotenv==0.21.0
supabase==0.5.8